(ciao_emacs) Fixed bug when running LPdoc from Emacs on files with same name
The temporary directory used when running LPdoc from Emacs was being
reused if subsequent files being documented had the same name, even if
they were at different locations. This resulted in the documetation
being (re)generated for the old file. Fixed by associating the
temporary directory with the full path of the file.

Src-commit: ea8486e2cb97ba2ab610a1f4d672507c8e64d7c1